% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mzml.R
\name{convert_mzml}
\alias{convert_mzml}
\title{Convert MRM-MS files to LC-MS style files}
\usage{
convert_mzml(mzML)
}
\arguments{
\item{mzML}{a character stirng of the absolute file path of a MRM-MS \code{.mzML} file}

\item{return}{logical; if \code{TRUE} then the `peaks` list is returned from the function. \code{Default is FALSE}.}
}
\description{
Convert MRM-MS \code{.mzML} files which contain a series of time-intensity pairs (chromatograms) to a series of LC-MS scan style \code{m/z}-intensity pairs.
}
\details{
Once the data is represented as a matrix of \code{m/z} and intensity pairs for each scan; the data can be further processed by open source tools such as \code{xcms}.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
