% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_mzml.R
\name{save_mzml}
\alias{save_mzml}
\title{Save Converted mzML}
\usage{
save_mzml(input, filename)
}
\arguments{
\item{input}{a list of \code{peaks} and \code{header} which has been created using the \code{convert_mzml} function}

\item{filename}{a character string of the filename to save converted files to.}
}
\description{
Save Converted mzML
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
