% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert MRM-MS files to LC-MS style files}
\usage{
convert(input, outpath)
}
\arguments{
\item{input}{a character stirng of the absolute file path of a MRM-MS \code{.mzML} file}

\item{outpath}{a character string of the location to save converted files to}
}
\description{
Convert MRM-MS \code{.mzML} files which contain a series of time-intensity pairs (chromatograms) to a series of LC-MS scan style \code{m/z}-intensity pairs.
}
\details{
Once the data is represented as a matrix of \code{m/z} and intensity for each scan; the data can be further processed by open source tools such as \code{xcms}.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
