% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mzml.R
\name{convert_mzml}
\alias{convert_mzml}
\title{Convert MRM-MS files to LC-MS style files}
\usage{
convert_mzml(mzML)
}
\arguments{
\item{mzML}{a character string of the absolute file path of a SRM/MRM-MS \code{.mzML} file}
}
\value{
a list of two elements
\itemize{
\item \strong{peaks} a list of spectra (@seealso \link{mzR(peaks)})
\item \strong{header} a \code{data.frame} of spectra header data (@seealso \link{mzR(header)})
}
}
\description{
Data aquired using Liquid Chromatography Triple Quadrupole Mass Spectromety (LC-QqQ-MS) operating in Selective Reaction Monitoiring (SRM) or Multiple Reaction Monitoring (MRM) mode
is stored in \code{mzML} files as a series of chromatrograms (retention time vs intensity). \code{convert_mzml} converts \code{mzML} chromatogram data into spectra; which is how data acquired using
LC-MS is stored. Once SRM/MRM data is represented as a series of spectra (\emph{m/z} vs intensity) then tools such as \code{xcms} can be used to process and analyse the data.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
